修改基于value分片的客户端

+++ redis-client/src/main/java/zx/soft/redis/client/hash/ManagerHash.java
@@ -16,6 +16,6 @@ public interface ManagerHash {
 
 	public List<String> getFields();
 
-	public void delFields(int timeSpan);
+	public void delFields(long timeSpan);
 
 }
